+++ controller/src/main/java/com/linbit/linstor/core/apicallhandler/controller/internal/RscDfnInternalCallHandler.java
@@ -7,6 +7,7 @@ import com.linbit.linstor.InternalApiConsts;
 import com.linbit.linstor.annotation.ApiContext;
 import com.linbit.linstor.annotation.PeerContext;
 import com.linbit.linstor.api.ApiCallRc;
+import com.linbit.linstor.api.ApiConsts;
 import com.linbit.linstor.api.interfaces.serializer.CtrlStltSerializer;
 import com.linbit.linstor.core.CoreModule;
 import com.linbit.linstor.core.apicallhandler.ScopeRunner;
@@ -22,6 +23,7 @@ import com.linbit.linstor.core.identifier.VolumeNumber;
 import com.linbit.linstor.core.objects.Resource;
 import com.linbit.linstor.core.objects.ResourceDefinition;
 import com.linbit.linstor.core.objects.Volume;
+import com.linbit.linstor.core.objects.VolumeDefinition;
 import com.linbit.linstor.dbdrivers.DatabaseException;
 import com.linbit.linstor.layer.resource.CtrlRscLayerDataFactory;
 import com.linbit.linstor.logging.ErrorReporter;
@@ -31,8 +33,12 @@ import com.linbit.linstor.propscon.InvalidValueException;
 import com.linbit.linstor.propscon.Props;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.security.AccessDeniedException;
+import com.linbit.linstor.storage.data.adapter.luks.LuksRscData;
+import com.linbit.linstor.storage.kinds.DeviceLayerKind;
+import com.linbit.linstor.utils.layer.LayerRscUtils;
 import com.linbit.locks.LockGuard;
 import com.linbit.locks.LockGuardFactory;
+import com.linbit.utils.Base64;
 
 import static com.linbit.linstor.core.apicallhandler.controller.internal.CtrlSatelliteUpdateCaller.notConnectedError;
 
@@ -40,7 +46,9 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Provider;
 
+import java.util.Collection;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.locks.ReadWriteLock;
@@ -218,6 +226,33 @@ public class RscDfnInternalCallHandler
             });
     }
 
+    private void modifyEncryptionPassword(Collection<Resource> resources) throws AccessDeniedException
+    {
+        for (Resource rsc : resources)
+        {
+            var luksRscLayerSet = LayerRscUtils.getRscDataByLayer(rsc.getLayerData(apiCtx), DeviceLayerKind.LUKS);
+
+            Iterator<VolumeDefinition> itVlmDfn = rsc.getResourceDefinition().iterateVolumeDfn(apiCtx);
+            while (itVlmDfn.hasNext())
+            {
+                VolumeDefinition vlmDfn = itVlmDfn.next();
+                String passphrase = vlmDfn.getProps(apiCtx).getProp(
+                    ApiConsts.KEY_PASSPHRASE, ApiConsts.NAMESPC_ENCRYPTION);
+                if (passphrase != null)
+                {
+                    errorReporter.logInfo(
+                        "Encryption clone modify password for " + rsc.getResourceDefinition().getName());
+                    for (var rscLayer : luksRscLayerSet)
+                    {
+                        LuksRscData<Resource> luksRscData = (LuksRscData<Resource>) rscLayer;
+                        var vlmLuksLayer = luksRscData.getVlmLayerObjects().get(vlmDfn.getVolumeNumber());
+                        vlmLuksLayer.setModifyPassword(Base64.decode(passphrase));
+                    }
+                }
+            }
+        }
+    }
+
     public Flux<ApiCallRc> handleCloneUpdateInTransaction(String rscName, int vlmNr, boolean success) {
         Flux<ApiCallRc> flux = Flux.empty();
         Peer currentPeer = peer.get();
@@ -257,9 +292,10 @@ public class RscDfnInternalCallHandler
 
             // if failed never finish
             boolean allCloned = !rscDfn.getFlags().isSet(apiCtx, ResourceDefinition.Flags.FAILED);
+            final Collection<Resource> resources = rscDfn.streamResource(apiCtx).collect(Collectors.toList());
             if (allCloned)
             {
-                for (Resource rsc : rscDfn.streamResource(apiCtx).collect(Collectors.toList()))
+                for (Resource rsc : resources)
                 {
                     for (Volume vlm : rsc.streamVolumes().collect(Collectors.toList()))
                     {
@@ -278,7 +314,8 @@ public class RscDfnInternalCallHandler
 
             if (allCloned)
             {
-                for (Resource rsc : rscDfn.streamResource(apiCtx).collect(Collectors.toList()))
+                modifyEncryptionPassword(resources);
+                for (Resource rsc : resources)
                 {
                     for (Volume vlm : rsc.streamVolumes().collect(Collectors.toList()))
                     {